app: use wxApp as entry point
