+++ src/aocmultiny/main.cpp
@@ -1,7 +1,7 @@
 #include "main.hpp"
 #include "cli/CLI.hpp"
 #include "irc/IRC.hpp"
-#include "gui/App.hpp"
+#include "gui/MainFrame.hpp"
 #include <dplib/DPAddress.hpp>
 #include <dplib/DPLobby.hpp>
 #include <dplib/DPGameAoC.hpp>
@@ -13,10 +13,9 @@
 
 using namespace std;
 using dplib::DPLobby;
+using aocmultiny::gui::MainFrame;
 
 namespace aocmultiny {
-
-const GUID DPSPGUID_NICE = { 0xe2dd8ebe, 0x1f03, 0x43b7, { 0x8d, 0x92, 0x9c, 0x6c, 0x2f, 0x5c, 0x44, 0x26 } };
 
 static wstring stow (string s) {
   wstring w;
@@ -26,11 +25,15 @@ static wstring stow (string s) {
   return w;
 }
 
-void main (vector<string> params) {
-  CoInitialize(NULL);
-  wcout << "[main] Starting" << endl;
+const GUID DPSPGUID_NICE = { 0xe2dd8ebe, 0x1f03, 0x43b7, { 0x8d, 0x92, 0x9c, 0x6c, 0x2f, 0x5c, 0x44, 0x26 } };
 
-  auto action = params.size() > 0 ? params[0] : "";
+bool App::OnInit () {
+  CoInitialize(NULL);
+
+  wcout << "[main] Starting" << endl;
+  vector<string> params = split(GetCommandLineA(), ' ');
+
+  auto action = params.size() > 1 ? params[1] : "";
   if (action == "host") {
     wcout << "[main] Hosting" << endl;
     auto game = new dplib::DPGameAoC();
@@ -39,8 +42,9 @@ void main (vector<string> params) {
     session
       ->setPlayerName("Host")
       ->launch();
+    return false;
   } else if (action == "join") {
-    GUID sessionId; IIDFromString(stow(params[1]).c_str(), &sessionId);
+    GUID sessionId; IIDFromString(stow(params[2]).c_str(), &sessionId);
     wcout << "[main] Joining " << to_wstring(sessionId) << endl;
     auto game = new dplib::DPGameAoC();
     auto address = new dplib::DPAddress(DPSPGUID_NICE);
@@ -49,29 +53,45 @@ void main (vector<string> params) {
     session
       ->setPlayerName("Rando")
       ->launch();
+    return false;
   } else if (action == "cli") {
     wcout << "[main] IRC" << endl;
     auto irc = new irc::IRC("localhost");
     auto cli = new cli::CLI(irc);
     cli->start();
     delete irc;
-  } else {
-    wxEntry(0, nullptr);
+    return false;
   }
 
-  wcout << "[main] Exiting" << endl;
+  if (!wxApp::OnInit()) {
+    return false;
+  }
+
+  wxTextEntryDialog dialog (NULL, wxT("Username"));
+  if (dialog.ShowModal() == wxID_CANCEL) {
+    return false;
+  }
+
+  auto username = dialog.GetValue().ToStdString();
+  this->irc = new irc::IRC("localhost");
+  auto frame = new MainFrame("AoCMulTiny", this->irc);
+  this->irc->nick(username);
+  this->irc->user(username);
+  frame->Show(true);
+  this->irc->join("#lobby");
+
+  return true;
+}
+
+int App::OnExit () {
+  delete this->irc;
 
   CoUninitialize();
-}
 
-}
-
-wxIMPLEMENT_APP_NO_MAIN(aocmultiny::gui::App);
-IMPLEMENT_WX_THEME_SUPPORT;
-
-int WINAPI WinMain (HINSTANCE instance, HINSTANCE prevInstance, LPSTR cmdLine, int showCmd) {
-  vector<string> params = split(cmdLine, ' ');
-
-  aocmultiny::main(params);
   return 0;
 }
+
+}
+
+wxIMPLEMENT_APP(aocmultiny::App);
+IMPLEMENT_WX_THEME_SUPPORT;
