+++ src/aocmultiny/main.hpp
@@ -1,5 +1,16 @@
+#pragma once
+#include "irc/IRC.hpp"
+#include <wx/wx.h>
+
+using aocmultiny::irc::IRC;
+
 namespace aocmultiny {
 
-void main ();
+class App: public wxApp {
+public:
+  IRC* irc;
+  virtual bool OnInit ();
+  virtual int OnExit ();
+};
 
 }
